Add files via upload
Using the electric vehicle dataset, create a model to forecast future EV adoption. For example, predict the number of electric vehicles in upcoming years based on the trends in the data. Build a regression model that forecasts future EV adoption demand based on historical trends in EV growth, types of vehicles, and regional data.